moved register to utils to avoid circular dependency
